serial: simplify baud rate divisor computation, fix fraction overflow
The formula to compute USARTDIV (when OVER8 is 0) is:

    USARTDIV = round (f_CK / (16 * baudrate))

As USARTDIV is a fixed point number with 4 bits fractionnal part (when
OVER8 is 0), this simplifies to:

    USARTDIV_reg = round (f_CK / baudrate)

The original code was computing the rounded fractionnal part which could
overflow. In this case, it could cause an error of 1 when the mantisa is
odd.