Add delay impls for TIM2/TIM5
Sometimes SYST cannot be used for delays, for example in the case of
RTIC where SYST is already used by the scheduler. For such cases, this
commit adds `DelayMs` and `DelayUs` trait impls (both for u32 and u16)
based on the 32-bit general purpose TIM2/TIM5 timers.

The implementation uses one-pulse mode, so that the timer doesn't need
to be explicitly stopped: The CEN flag is automatically cleared once the
auto-reload value in the ARR is reached.

For backwards compatibility, the old `Delay` struct was _not_ renamed to
`SystickDelay`, but that might be something to consider.