Add PWM channels for timers 9-14
Adds support for PWM outputs to timers 9-14 as mentioned in issue #361.

Eliminates the "tim8" configuration feature block in favor of keeping all feature gates based on the part number.

Uncomments two pins for TIM5 that are on the datasheet, and I don't see an errata suggesting they shouldn't be used.