update to `embedded-hal=1.0.0-alpha.10`
there were various breaking changes in alpha.9 and alpha.10 which have
been applied here. see [their changelog][embedded-hal-1.0.0-alpha.10]
for further details.

important: this is completely untested beyond "it compiles", please test
this carefully!
compilation has been tested with the `stm32f401` feature, the CI will
show whether it'll also compile for all others.

note that this does only the bare minimum to make it build against the
new version - it does not implement any of the newly added APIs. this
can be done in subsequent commits.

this fixes #602

[embedded-hal-1.0.0-alpha.10]: https://github.com/rust-embedded/embedded-hal/blob/v1.0.0-alpha.10/embedded-hal/CHANGELOG.md