update to `embedded-hal` v1.0.0 (no more RC!)
it finally happened: the stable release of `embedded-hal` has just been
released.

note that between RC.3 and the final release the `ToggleableOutputPin`
has been removed. the `toggle` method was moved to `StatefulOutputPin`
which does the same as the HAL-internal `toggle` method, thus no
implementation is needed anymore on our side.