Prevent starting a new I2C transaction before previous stop finishes.
When a stop condition is generated by the I2C DMA callback, the code
does not wait for the finish of the stop condition. If while the stop
condition is being generated and a new I2C transaction attempts to
start, the new transaction will be inadvertently terminated by the
previous stop condition.

Since ISRs should finish as quickly as possible, the ISRs should not
spin-wait until the stop condition finishes. Instead, the spin-wait
should be placed before starting a new transaction.

The patch also removes incorrect comments that falsely indicate that
the code will spin-wait until the stop condition finishes.